% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_landcover.R
\name{classify_landcover}
\alias{classify_landcover}
\title{Classify landcover}
\usage{
classify_landcover(x, ...)
}
\arguments{
\item{x}{object of class sf, sfc, sfg, or SpatRaster}

\item{...}{Unused}
}
\description{
Classify landcover
}
