% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_landcover.R
\name{classify_landcover.sf}
\alias{classify_landcover.sf}
\title{Classify landcover data for simple features object}
\usage{
\method{classify_landcover}{sf}(x, source = "LSPT", ...)
}
\arguments{
\item{x}{object of class sf, sfc or sfg}

\item{source}{character string indicating the source of the land cover
polygons; currently only "LSPT" is supported}

\item{...}{Unused}
}
\description{
Prepare a land cover data set for use with the other functions in this
package by aligning the land cover classifications with those required by the
\link{metrics} data as provided in the \link{key} and corresponding predictors required
for fitting species distribution models.
}
\details{
This function provides support for cross-walking polygon-based vegetation
data, such as the "Habitat_types_modern_forLSPT" layer published in the Landscape
Scenario Planning Tool v3 (LSPT), to the land cover classes required for use
with this package. Ultimately, this function may be updated to extend support
to other land cover data sources, but it currently relies on the fields
provided in that layer, especially: \code{Habitat_Type}, \code{Source_classification},
\code{Crop2016}, and, \code{Tidal}.

The function returns an sf object with additional fields:
\itemize{
\item \code{CODE_NAME}: most specific land cover class designation, matching those provided in the \link{key}
\item \code{CODE_NUM}: corresponding code number, matching those in the \link{key}, for creating rasters
\item \code{PREDICTOR_RIPARIAN}: corresponding predictor name used in the riparian landbird SDMs
\item \code{PREDICTOR_WATERBIRD_FALL}: corresponding predictor name used in the waterbird SDMs for the fall season
\item \code{PREDICTOR_WATERBIRD_WIN}: corresponding predictor name used in the waterbird SDMs for the winter season
\item \code{PREDICTOR_TIMA}: corresponding predictor name used in the tidal marsh bird SDMs
}

The \code{CODE_NAME} field can be rasterized for further analysis with the
\link{metrics} data and converted to predictors for fitting the SDMs by calling
this function again on the raster (see \code{\link[=classify_landcover.SpatRaster]{classify_landcover.SpatRaster()}}.
Alternatively, each predictor field can be rasterized directly from this
output. The results should be mostly identical, except for a few special
cases for the riparian predictors; see \link{predictors_riparian} for more
information.
}
\seealso{
\link{key}, \link{predictors_riparian}, \link{predictors_waterbird_win},
\link{predictors_waterbird_fall}, \link{predictors_tima}
}
