% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-py.R
\docType{data}
\name{.py_state}
\alias{.py_state}
\title{Python helpers}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.py_state
}
\description{
Internal helpers that perform one-time Python initialization and provide
per-module lazy loaders. These functions are internal and should not be
called directly by users.
}
\keyword{internal}
