% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_tima_patchsize.R
\name{estimate_tima_patchsize}
\alias{estimate_tima_patchsize}
\title{Estimate tidal marsh patch size}
\usage{
estimate_tima_patchsize(
  x,
  directions = 8,
  zeroAsNA = TRUE,
  fill = FALSE,
  dir = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{x}{SpatRaster or list of SpatRasters; see Details}

\item{directions}{integer passed to \code{\link[terra:patches]{terra::patches()}} indicating which cells
are considered adjacent. Should be 8 (Queen's case) or 4 (Rook's case)}

\item{zeroAsNA}{logical passed to \code{\link[terra:patches]{terra::patches()}}. If TRUE treat cells
that are zero as if they were NA}

\item{fill}{logical. If TRUE replaces all non-tidal wetland vegetation with
0.}

\item{dir}{Optional string representing directory passed to
\code{\link[terra:writeRaster]{terra::writeRaster()}}, as (\code{dir/SDM/landscape_name}). See Details.}

\item{overwrite}{logical. If \code{TRUE}, output is overwritten}

\item{...}{additional arguments passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}}
}
\value{
SpatRaster
}
\description{
Additional preparation required prior to running focal statistics on a
landscape raster via Python, to generate tidal wetland patch size estimates
for use with tidal marsh bird ('tima') models.
}
\details{
The input should be a SpatRaster or list of SpatRasters resulting from
running \code{\link[=python_focal_prep]{python_focal_prep()}} with \code{SDM = "tima"}. This function extracts the
\code{TWET} predictor layer, representing all tidal wetland vegetation, identifies
distinct contiguous patches, and assigns each pixel within each patch a value
corresponding to the count of pixels within the patch.
}
\examples{
codenums = DeltaMultipleBenefits::key$CODE_NUM
r <- terra::rast(matrix(sample(codenums, size = 1000, replace = TRUE), ncol = 100, nrow = 100))
tima_pred = python_focal_prep(r, SDM = 'tima')
tima_psize = estimate_tima_patchsize(tima_pred)

}
