% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_tima.R
\docType{data}
\name{predictors_tima}
\alias{predictors_tima}
\title{Predictors used in distribution models for tidal marsh bird focal species}
\format{
\subsection{\code{key} A data frame with 81 rows and 9 columns:}{

\describe{
\item{CODE_NUM}{Numeric value used to encode rasters, matching \code{key}}
\item{CODE_NAME}{Text string joining major land cover classes to subclasses with a '_', matching \code{key}}
\item{PREDICTOR_NAME}{Corresponding predictor name in tidal marsh bird distribution models}
\item{PREDICTOR_NUM}{Corresponding numeric value used to reclassify land cover rasters for use with tidal marsh bird distribution models}
\item{WETL}{Additional predictor grouping wetland subclasses together; 1 indicates the included subclasses}
\item{NWET}{Additional predictor grouping nontidal wetland subclasses together; 1 indicates the included subclasses}
\item{TWET}{Additional predictor grouping tidal wetland subclasses together; 1 indicates the include subclasses}
\item{RFOR}{Additional predictor grouping riparian forest subclasses together; 1 indicates the included subclasses}
\item{RSCR}{Additional predictor grouping riparian scrub subclasses together; 1 indicates the included subclasses}
}
}
}
\source{
Dybala et al. In review
}
\usage{
predictors_tima
}
\description{
Predictor names and how they crosswalk to the major land cover classes and
subclasses listed in the vegetation \code{key}. Land cover classes or subclasses
with \code{NA} values in \code{PREDICTOR_NAME} and \code{PREDICTOR_NUM} are excluded from
these models. This data set is primarily for use in the
\code{\link[=classify_landcover]{classify_landcover()}} function.
}
\keyword{datasets}
