% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_waterbirds_win.R
\docType{data}
\name{predictors_waterbirds_win}
\alias{predictors_waterbirds_win}
\title{Predictors used in distribution models for waterbird groups during the winter
season}
\format{
\subsection{\code{key} A data frame with 81 rows and 4 columns:}{

\describe{
\item{CODE_NUM}{Numeric value used to encode rasters, matching \code{key}}
\item{NAME_FULL}{Text string joining major land cover classes to subclasses with a '_', matching \code{key}}
\item{PREDICTOR_NAME}{Corresponding predictor name in tidal marsh bird distribution models}
\item{PREDICTOR_NUM}{Corresponding numeric value used to reclassify land cover rasters for use with tidal marsh bird distribution models}
}
}
}
\source{
Dybala et al. 2023 (https://doi.org/10.15447/sfews.2023v21iss3art4)
}
\usage{
predictors_waterbirds_win
}
\description{
Predictor names and how they crosswalk to the major land cover classes and
subclasses listed in the vegetation \code{key}. Land cover classes or subclasses
with \code{NA} values in \code{PREDICTOR_NAME} and \code{PREDICTOR_NUM} are excluded from
these models. This data set is primarily for use in the
\code{\link[=reclassify_landcover]{reclassify_landcover()}} function. This crosswalk is nearly identical to
\code{predictors_waterbirds_fall} except that in the winter, agricultural fields
specified as "winter wheat" are separated from other grains.
}
\keyword{datasets}
