% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify_landcover.R
\name{reclassify_landcover}
\alias{reclassify_landcover}
\title{Reclassify landscape rasters by SDM}
\usage{
reclassify_landcover(landscape_stack, SDM)
}
\arguments{
\item{landscape_stack}{SpatRaster created by \code{\link[terra:rast]{terra::rast()}}}

\item{SDM}{The name of intended species distribution model: \code{"riparian"},
\code{"waterbird_fall"}, \code{"waterbird_win"}, or \code{"tima"}}
}
\value{
SpatRaster with layers for each land cover class
}
\description{
Prepare for fitting SDMs by reclassifying a landscape rasters according to
the classifications used by a specific species distribution model (SDM).
}
\details{
This function is called by \code{\link[=python_focal_prep]{python_focal_prep()}} on a set of landscape
rasters generated by segregating an input landscape by class. It is not
intended to be called directly.
}
\examples{
#r <- terra::rast(matrix(sample(c(0:4), size = 100, replace = TRUE),
#         ncol = 10, nrow = 10)
#levels(r) <- c('RIPARIAN_FOREST_POFR', 'RIPARIAN_FOREST_QULO', 'WETLAND', 'ORCHARD', 'VINEYARD')
#layernames = terra::freq(r)$label
#s = terra::segregate(r, other = 0)
#names(s) = terra::freq(r)$label
#example = reclassify_landcover(s, SDM = 'RIPARIAN')
}
