% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roosts_original.R
\docType{data}
\name{roosts_original}
\alias{roosts_original}
\title{Traditional nighttime crane roost locations}
\format{
\subsection{\code{roosts_original} A SpatVector containing with 94 geometries 1}{

attribute field:
\describe{
\item{Roost_ID}{Numeric value used to identify distinct roosts}
}
}
}
\usage{
roosts_original
}
\description{
Original crane roost locations used in generating the "distance to roost"
predictor variable used in the crane distribution models within the
\code{DeltaMultipleBenefits} framework. The assumed future locations of roosts
under scenarios of landscape change may require modification if land covers
become incompatible (see \code{\link[=update_roosts]{update_roosts()}})
}
\keyword{datasets}
