% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_change.R
\name{sum_change}
\alias{sum_change}
\title{Summarize net change between scenario and baseline landscapes}
\usage{
sum_change(scoredat)
}
\arguments{
\item{scoredat}{tibble; see Details}
}
\value{
tibble
}
\description{
Calculate the difference in the landscape-level metric totals between a
baseline landscape and one or more scenario landscapes
}
\details{
This function takes a tibble containing the metric scores for
multiple landscapes, including those produced by \code{\link[=sum_habitat]{sum_habitat()}} or
\code{\link[=sum_metrics]{sum_metrics()}}, with or without scores by ZONE within the entire
landscape. Scores for a baseline landscape are subtracted from scores for
one or more scenario landscapes to calculate a net change and the
proportion and percent change from baseline, along with propagation of
uncertainty.

This function expects \code{scoredat} to contain the following fields, in
alignment with the output of \code{\link[=sum_habitat]{sum_habitat()}} and \code{\link[=sum_metrics]{sum_metrics()}}:
\itemize{
\item scenario: a character field used to identify the name of the landscape
being examined; at least one should be called \code{'baseline'}
\item SCORE_TOTAL, SCORE_TOTAL_SE: numeric fields representing the total
landscape-level values for each metric and associated uncertainty
}
}
\examples{
# See vignette
}
\seealso{
\code{\link[=sum_habitat]{sum_habitat()}}, \code{\link[=sum_metrics]{sum_metrics()}}
}
