% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_habitat.R
\name{sum_habitat}
\alias{sum_habitat}
\title{Summarize total habitat scores}
\usage{
sum_habitat(
  pathin,
  SDM = NULL,
  landscape_name = NULL,
  regex = ".tif$",
  zones = NULL,
  subtype = NULL,
  rollup = TRUE,
  key = NULL,
  scale = NULL
)
}
\arguments{
\item{pathin}{Character string defining the filepath to the highest-level
directory containing the predicted presence/absence or probability of
presence from species distribution models, such as those created from
running \code{\link[=fit_SDM]{fit_SDM()}} or \code{\link[=transform_SDM]{transform_SDM()}}}

\item{SDM, landscape_name}{Optional character strings defining the
subdirectories within \code{pathin} for which habitat should be summarized; see
Details.}

\item{regex}{Passed to \code{list.files} for selecting a subset of rasters from
\code{pathin/SDM/landscape_name}; default is ".tif$"}

\item{zones}{Optional \code{SpatRaster} or character string giving the filepath to
a raster encoding zones within which pixel values should be summarized}

\item{subtype}{Optional character string appended to the field
METRIC_SUBTYPE, such as for distinguishing probability of presence from
presence/absence}

\item{rollup}{Logical; If \code{TRUE} (default), summarize total habitat across
all species/groups by set of SDMs and add to output}

\item{key}{Optional tibble, dataframe, or character string defining the
filepath passed to \code{\link[readr:read_delim]{readr::read_csv()}}, used to translate the individual
species names as encoded in the file names in pathin to readable METRIC
names in the output table; see Details}

\item{scale}{Optional value by which to scale the results; see Details}
}
\value{
tibble
}
\description{
Calculate the sum of the predicted presence/absence or probability of
presence from species distribution models corresponding to a set of landscape
rasters.
}
\details{
By default, this function will calculate the sum of all pixel values
in each raster found in \code{pathin} and any subdirectories, to efficiently
process multiple rasters located in the \code{pathin} directory at once. The
file structure within \code{pathin} is used to infer the name of the
corresponding \code{SDM} and \code{landscape_name}. However, \code{SDM} or both \code{SDM} and
\code{landscape_name} can be optionally specified to only process a subset of
these. The \code{regex} option can also be used to specify a subset of the
rasters within \code{pathin/SDM/landscape_name}.

If SDM is one of the recognized models (\code{"riparian"}, \code{"waterbird_fall"},
\code{"waterbird_win"}, or \code{"tima"}), the output will produce nicer labels for
the \code{METRIC} and \code{METRIC_SUBTYPE} fields. Otherwise, \code{METRIC_SUBTYPE} will be
filled with the \code{SDM} value.

If provided, \code{key} should refer to a tibble, dataframe, or filepath to a
CSV containing the fields \code{spp} and \code{label}, used for converting the
filenames of the rasters within \code{pathin} to a more readable label, which
will be renamed \code{METRIC} in the final output.

If provided, the sum of all pixel values will be multiplied by \code{scale}. For
example, to rescale the total in terms of the total area, enter the area of
each pixel.
}
\examples{
# See vignette
}
\seealso{
\code{\link[=fit_SDM]{fit_SDM()}}, \code{\link[=transform_SDM]{transform_SDM()}}
}
