% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_landcover.R
\name{sum_landcover}
\alias{sum_landcover}
\title{Summarize total area of land cover classes}
\usage{
sum_landcover(
  landscapes,
  mask = NULL,
  zones = NULL,
  pixel_area = 1,
  rollup = TRUE
)
}
\arguments{
\item{landscapes}{Either a \code{SpatRaster} with one or more layers, or a character
string giving the filepath to a directory containing one or more raster
landscapes (e.g. scenarios) to be summarized.}

\item{mask}{Optional \code{SpatRaster} or character string giving the filepath to
a raster that should be used to mask the output, e.g. a study area boundary}

\item{zones}{Optional \code{SpatRaster} or character string giving the filepath to
a raster encoding zones within which pixel values should be summarized}

\item{pixel_area}{Numeric value representing the area of each pixel; hectares
preferred but any units may be provided; default is 1, resulting in a total
count}

\item{rollup}{Logical; whether or not riparian and managed wetland subclasses
should be rolled up into a total area of riparian and managed wetlands,
respectively. See Details.}
}
\value{
tibble
}
\description{
Summarize the area or number of pixels of each land cover class from a set of
landscape rasters
}
\details{
This function will summarize the frequency of pixels of each land
cover class in each landscape raster, either provided as a \code{SpatRaster}
created by \code{\link[terra:rast]{terra::rast()}} with one or more layers, or as a filepath to a
directory containing one or more \code{tif} raster files to be summarized. This
function is generally agnostic as to the names of the rasters and the land
cover classes it expects, but it does expect the rasters to have unique
names and to have land cover class levels defined (i.e., not just contain
numeric codes; see \code{\link[terra:factors]{terra::levels()}}). However, if \code{rollup = TRUE} (the
default), riparian and wetland subclasses are expected to have
names that begin with \code{'RIPARIAN_'} or \code{'WETLAND_'}. These
subclasses will be summarized into a total area of \code{'RIPARIAN'} and
\code{'WETLAND'} land cover classes.

This function relies on \code{\link[terra:freq]{terra::freq()}} to summarize the total number of
cells of each value within each raster and optionally each zone provided,
which are then multiplied by \code{pixel_area} as an estimate of the total area
of each land cover class in each landscape. The value for \code{pixel_area} can
be provided in any units you prefer, but if this function is intended to
produce \code{areadat} for use with \code{\link[=sum_metrics]{sum_metrics()}}, we recommend using hectares
to ensure alignment with the units in which metrics were estimated.

By default, the total area of each unique land cover class will be
summarized for the entire raster provided, but optionally, the area can be
subset by providing a \code{mask} and/or \code{zones}. A \code{mask} should be a single
\code{SpatRaster} or filepath to a \code{tif} raster file that will be passed to
\code{\link[terra:mask]{terra::mask()}} to exclude any areas with \code{NA} values. In addition, \code{zones}
should be a single \code{SpatRaster} or filepath to a \code{tif} raster file that
will be passed to \code{\link[terra:zonal]{terra::zonal()}}. Values in \code{zones} should define two or
more unique regions within which the area of each land cover class should
be summarized, and the names of each region should be defined (see
\code{\link[terra:factors]{terra::levels()}}).
}
\examples{
# See vignette
}
\seealso{
\code{\link[=sum_habitat]{sum_habitat()}}, \code{\link[=sum_metrics]{sum_metrics()}}
}
