% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_change_SDM.R
\name{calc_change_SDM}
\alias{calc_change_SDM}
\title{Calculate the difference between predictions of species presence for a
baseline and scenario landscape}
\usage{
calc_change_SDM(baseline, scenario, differentiate = TRUE)
}
\arguments{
\item{baseline, scenario}{SpatRasters created by \code{\link[terra:rast]{terra::rast()}}; difference
will represent the change from the baseline to the scenario}
}
\description{
Calculate the difference between predictions of species presence for a
baseline and scenario landscape
}
\details{
This function is designed to handle the predictions for multiple
species provided as layers in \code{baseline} and \code{scenario} SpatRasters,
matched by layernames. The difference is calculated as the predicted value
for the scenario minus the predicted value for the baseline, such that
positive values in the result represent an increased probability of
presence, and negative values represent a reduced probability of presence.

If \code{differentiate = TRUE}, the function is also designed to distinguish
between two types of zero values in the result: locations where the
predicted value is the same in both landscapes (which will retain a value
of zero difference), and locations where the predicted value is zero in
both landscapes (which will be converted to NA).
}
