% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_landcover.R
\name{classify_landcover.SpatRaster}
\alias{classify_landcover.SpatRaster}
\title{Classify landcover for SpatRaster object}
\usage{
\method{classify_landcover}{SpatRaster}(x, SDM, coltab = TRUE, ...)
}
\arguments{
\item{x}{SpatRaster}

\item{SDM}{The name of intended species distribution model: \code{"riparian"},
\code{"waterbird_fall"}, \code{"waterbird_win"}, or \code{"tima"}}

\item{coltab}{logical; if TRUE add default color palette}

\item{...}{Unused}
}
\description{
Prepare for fitting SDMs by reclassifying an existing landscape raster
according to the classifications used by a specific set of species
distribution models (SDM).
}
\details{
Calls on internal datasets to crosswalk from land cover classes listed in the
\link{key} to the predictors expected by the selected SDM group. The input raster
should already be encoded with the land cover classes listed in the \link{key}. To
help with creating such a raster, see \code{\link[=classify_landcover.sf]{classify_landcover.sf()}} to map land
cover polygons to the land cover classes in the \link{key}.

A warning is given if there are land cover classes present in the landscape
that do not map to any of the predictors for the selected SDM group, or if
there are land cover classes missing from the landscape that are expected by
the selected SDM group. These warnings may represent significant problems for
fitting SDMs and should be carefully reviewed.  In either case, it is
recommended to review the corresponding internal datasets
(\link{predictors_riparian}, \link{predictors_waterbird_fall},
\link{predictors_waterbird_win}, or \link{predictors_tima}) for the list of expected
predictors and how they map to land cover classes in the \link{key}. Check whether
the selected SDM group expects more specific land cover classes or
subclasses; the input raster may need to be reclassified before proceeding.
}
\examples{
r <- terra::rast(matrix(sample(c(11,19,71,72,90), size = 100, replace = TRUE),
         ncol = 10, nrow = 10))
r <- suppressWarnings(classify_landcover(r, SDM = 'riparian'))
}
