% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_directory.R
\name{create_directory}
\alias{create_directory}
\title{create_directory}
\usage{
create_directory(filepath)
}
\arguments{
\item{filepath}{character string representing a file path}
}
\value{
Prints either "Creating directory: filepath" if directory did not
already exist, or "Writing to directory: filepath" if it did already exist.
}
\description{
Helper function called within other functions to check whether the intended
output directory exists, and if not, create it. Particularly useful when many
files are generated programmatically with an expected structure.
}
\examples{
\dontrun{
create_directory(filepath = "Desktop/example")

# will create nested directories as needed
create_directory(filepath = "Desktop/example/test")
}

}
