% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_predictor_stack.R
\name{create_predictor_stack}
\alias{create_predictor_stack}
\title{Create raster stack representing SDM predictors}
\usage{
create_predictor_stack(x, SDM, fill = TRUE)
}
\arguments{
\item{x}{SpatRaster; can only have 1 layer}

\item{SDM}{The name of intended species distribution model: \code{"riparian"},
\code{"waterbird_fall"}, \code{"waterbird_win"}, or \code{"tima"}}

\item{fill}{logical; see Details}
}
\value{
SpatRaster with separate layers for each land cover class included
as a predictor in the selected SDM representing the presence (1) and
absence (0)
}
\description{
Prepare for fitting SDMs by transforming a landscape raster into a stack of
rasters representing the required predictors.
}
\details{
This function is called by \code{\link[=python_focal_prep]{python_focal_prep()}} and is not intended to be
called directly. Segregates a landscape raster into separate layers
representing each land cover class. Also calls on internal datsets to create
broader grouping variables as required for the selected SDM. The input raster
should already be classified according to the land cover classifications
expected by the selected SDM. See \code{\link[=classify_landcover.SpatRaster]{classify_landcover.SpatRaster()}}.

A warning is given if land cover classes expected by the model are absent
from the landscape. If fill = TRUE (the defualt), the function will create
rasters with all zero values for each missing land cover class. However, the
input raster should be carefully reviewed to ensure they are truly absent and
have not been excluded from the landscape unintentionally. If needed, the
resulting layers can be replaced manually before proceeding with
\code{\link[=python_focal_run]{python_focal_run()}}.
}
\examples{
r <- terra::rast(matrix(sample(c(11,19,71,72,90), size = 100, replace = TRUE),
         ncol = 10, nrow = 10))
r = suppressWarnings(create_predictor_stack(r, SDM = 'riparian'))
}
\seealso{
\code{\link[=classify_landcover]{classify_landcover()}}
}
