% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_predictor_stack.R
\name{create_predictor_stack}
\alias{create_predictor_stack}
\title{Create raster stack representing SDM predictors}
\usage{
create_predictor_stack(x, SDM, classified = FALSE, fill = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{SpatRaster; can only have 1 layer}

\item{SDM}{The name of intended species distribution model: \code{"riparian"},
\code{"waterbird_fall"}, \code{"waterbird_win"}, or \code{"tima"}}

\item{classified}{logical; see Details}

\item{fill}{logical; see Details}

\item{verbose}{logical; passed to \code{\link[=classify_landcover.SpatRaster]{classify_landcover.SpatRaster()}}}
}
\value{
SpatRaster with separate layers for each land cover class included
as a predictor in the selected SDM representing the presence (1) and
absence (0)
}
\description{
Prepare for fitting SDMs by transforming a landscape raster into a stack of
rasters representing the required predictors.
}
\details{
This function is called by \code{\link[=python_focal_prep]{python_focal_prep()}} and is not intended to be
called directly. Segregates a landscape raster into separate layers
representing each land cover class. Also calls on internal datsets to create
broader grouping variables as required for the selected SDM. The input raster
should already be encoded with the land cover classes listed in the \link{key}.

If \code{classified = FALSE} (the default), this function calls
\code{\link[=classify_landcover.SpatRaster]{classify_landcover.SpatRaster()}} to reclassify the input raster according to
the land cover classifications expected by the selected SDM. (See
documentation from that function for information on warning messages.) If
\code{classified = TRUE}, it will be assumed to already be classified correctly,
which may be of particular use for the 'riparian' models. See Vignette for
details.

A warning is given if land cover classes expected by the model are absent
from the provided landscape. In that case, if \code{fill = TRUE} (the default),
additional layers will be created with all zero values for each missing land
cover class. However, the input landscape should be carefully reviewed to
ensure they are truly absent and have not been excluded from the landscape
unintentionally. If needed, the resulting layers can be replaced manually
before proceeding with \code{\link[=python_focal_run]{python_focal_run()}}.
}
\examples{
r <- terra::rast(matrix(sample(c(11,19,71,72,90), size = 100, replace = TRUE),
         ncol = 10, nrow = 10))
r = suppressWarnings(create_predictor_stack(r, SDM = 'riparian'))
}
\seealso{
\code{\link[=classify_landcover]{classify_landcover()}}
}
