% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_riparian.R
\docType{data}
\name{predictors_riparian}
\alias{predictors_riparian}
\title{Predictors used in distribution models for riparian focal species}
\format{
\subsection{\code{predictors_riparian} A data frame with 81 rows and 7 columns:}{

\describe{
\item{CODE_NUM}{Numeric value used to encode rasters, matching \code{key}}
\item{CODE_NAME}{Text string joining major land cover classes to subclasses with a '_', matching \code{key}}
\item{PREDICTOR_NAME}{Corresponding predictor name in tidal marsh bird distribution models}
\item{PREDICTOR_NUM}{Corresponding numeric value used to reclassify land cover rasters for use with tidal marsh bird distribution models}
\item{RIPARIAN}{Additional predictor grouping riparian subclasses together; 1 indicates the included subclasses}
\item{WETLAND}{Additional predictor grouping wetland subclasses together; 1 indicates the included subclasses}
\item{NOTES}{See Details}
}
}
}
\source{
Dybala et al. 2023 (https://doi.org/10.15447/sfews.2023v21iss3art4)
}
\usage{
predictors_riparian
}
\description{
Predictor names and how they crosswalk to the major land cover classes and
subclasses listed in the vegetation \code{key}. Land cover classes with a \code{1} in
the \code{RIPARIAN} or \code{WETLAND} fields indicate they should be included in
additional predictors representing the proportion cover of those broader
vegetation classes. This data set is primarily for use in the
\code{\link[=classify_landcover]{classify_landcover()}} function.
}
\details{
The \code{NOTES} field provides additional information on vegetation classes that
were ultimately excluded from model predictions and can be safely ignored
(i.e. the \code{WOODLAND&SCRUB} class), as well as notes on more specific
vegetation types from the original source classifications that were excluded
from the predictor when these models were first developed.

Caution: For riparian landbird SDMs, use of \code{\link[=classify_landcover.SpatRaster]{classify_landcover.SpatRaster()}}
on a raster developed from the primary land cover classifications produced by
\code{\link[=classify_landcover.sf]{classify_landcover.sf()}} cannot be completely consistent with the original
classification scheme used to develop these models because of the minor
exceptions detailed in the \code{NOTES} field. We recommend instead creating a
raster directly based on the \code{PREDICTORS_RIPARIAN} field produced by
\code{\link[=classify_landcover.sf]{classify_landcover.sf()}}.
}
\keyword{datasets}
