% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_dist.R
\name{python_dist}
\alias{python_dist}
\title{Calculate Euclidean distance via Python}
\usage{
python_dist(
  x,
  scale = NULL,
  mask = NULL,
  python = NULL,
  dir = NULL,
  SDM = NULL,
  landscape_name = NULL,
  filename = "droost_km.tif",
  ...
)
}
\arguments{
\item{x}{Filepath or SpatRaster to be processed}

\item{scale}{Optional character string for scaling the results; See Details}

\item{mask}{Optional \code{SpatRaster} or character string giving the filepath to
a raster that should be used to mask the output, e.g. a study area boundary}

\item{python}{Optional filepath to the preferred version of arcpy, passed to
\code{reticulate::use_python}; See details.}

\item{dir, SDM, landscape_name}{Optional; Character strings defining the
filepath where output raster should be written (\code{dir/SDM/landscape_name})}

\item{filename}{name of the output raster, including file extension; default
is 'droost_km.tif', the name of the predictor required by the waterbird
models}

\item{...}{Additional arguments passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}}
}
\value{
SpatRaster
}
\description{
Wrapper function to call a python script for calculating Euclidean distances
on landscape rasters. On first call this function imports the arcpy module
and Spatial Analyst extensions then sources the Python script
"dist_stats.py".
}
\details{
This function calls the \code{dist_stats.py} script to calculate the
Euclidean distance for all cells in the input raster without a value to the
nearest cell with a value (e.g., for calculating distance to a crane roost
or a stream). Optionally, results can be scaled, masked, and written to
\code{dir/SDM/landscape_name/filename}.

Currently supported scale options include: \code{km} to divide the results by
1000 and return distances in kilometers or \code{sqrt} to take the square root
of the results.

The parameters \code{dir}, \code{SDM}, \code{landscape_name}, and \code{filename} are passed to
\code{\link[=file.path]{file.path()}} to contstruct filenames passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}. If
multiple values for one of these character strings is provided, the output
raster will be written to more than one directory, e.g. if needed as a
predictor for multiple SDMs or landscape names.

This function relies on the availability of \code{arcpy} and Spatial Analyst
extensions. An attempt will be made to load these the first time this
function (or \code{\link[=python_focal_stats]{python_focal_stats()}}) is called in each session, and by
default will look here: \verb{C:/Program Files/ArcGIS/Pro/bin/Python/envs/arcgispro-py3/python.exe}; use the
\code{python} argument to specify a different pathway.
}
\examples{
# See vignette
}
