% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_dist.R
\name{python_dist}
\alias{python_dist}
\title{Calculate Euclidean distance via Python}
\usage{
python_dist(
  pathin,
  landscape_name,
  copyto = NULL,
  pathout,
  filename = "droost_km.tif",
  scale = NULL,
  maskpath = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{pathin}{Filepath for the directory containing input rasters to be
processed, such as those created from running \link{python_focal_prep}}

\item{landscape_name}{Subdirectory containing the rasters for a target
landscape scenario for generating focal_stats}

\item{copyto}{Optional character string containing alternate \code{landscape_name}
to which the results of \code{\link[=python_dist]{python_dist()}} should be copied.}

\item{pathout}{filepath for the directory where output rasters should be
written}

\item{filename}{name of resulting layer, including file extension; default is
'droost_km.tif', the name of the predictor required by the waterbird models}

\item{scale}{Optional character string for scaling the results; See Details}

\item{maskpath}{Optional filepath to a raster that should be used to mask the
output, e.g. a study area boundary}

\item{overwrite}{Logical; passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}. Applies to
\code{copyto} only.}
}
\value{
Nothing; all files written to \code{pathout}
}
\description{
Function to call python script for calculating Euclidean distances on
landscape rasters via arcpy.
}
\details{
Calls the \link{dist_stats.py} function to calculate the Euclidean
distance for all cells in the input raster without a value to the nearest
cell with a value (e.g., for calculating distance to a crane roost or a
stream).

Resulting distances may be scaled using the \code{scale} argument. Currently
supported options include: \code{km} to divide the results by 1000 and return
distances in kilometers or \code{sqrt} to take the square root of the results.

Important: This function requires the availability of arcpy and Spatial
Analyst extensions. While these statistics can be entirely calculated in R,
arcpy is much faster. Note: the initial output of \link{dist_stats.py} will not
overwrite existing rasters; old versions must be deleted before re-running.
}
\examples{
# See vignette
}
\seealso{
\code{\link[=python_focal]{python_focal()}}
}
