% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_focal_finalize.R
\name{python_focal_finalize}
\alias{python_focal_finalize}
\title{Final processing of focal stats for SDMs}
\usage{
python_focal_finalize(
  pathin,
  scenario_name,
  SDM,
  scale,
  pathout,
  overwrite = FALSE,
  maskpath = NULL,
  cover = FALSE
)
}
\arguments{
\item{pathin}{Character string; Filepath to directory where source rasters
can be located; passed to \code{\link[terra:rast]{terra::rast()}}}

\item{scenario_name}{Character string; Name of the landscape scenario being
evaluated}

\item{SDM}{Character string; Name of intended species distribution model:
\code{"riparian"}, \code{"waterbird_fall"}, or \code{"waterbird_win"}.}

\item{scale}{Character string; Spatial scale over which focal stats are
summarized: \code{50}, \code{2000}, \code{5000}, or \code{10000}.}

\item{pathout}{Character string; Filepath to directory where output rasters
should be written; passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}}

\item{overwrite}{Logical; passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}}

\item{maskpath}{Optional filepath to a raster that should be used to mask the
output, e.g. a study area boundary}

\item{cover}{Logical; default is \code{FALSE}. If \code{TRUE}, \code{maskpath} must not be
NULL; See Details.}
}
\value{
Nothing returned to R environment. Writes rasters to \code{pathout} for
each land cover class.
}
\description{
Renames and rescales output from \code{\link[=python_focal]{python_focal()}} as needed to match expected
inputs for species distribution models (SDMs). Includes options to mask by
another raster and fill missing values with zero.
}
\details{
Function expects source files to be in a directory structure created
by \code{\link[=python_focal]{python_focal()}}, which is also used to inform the final processing
steps: \code{pathin/scenario_name/SDM/scale}. All .tif files in this source
directory will be read in, and optionally masked by the raster at
\code{maskpath}. If \code{cover = TRUE}, pixels in \code{maskpath} with a value of 1 will
also be replaced with a value of 0, and passed to \code{\link[terra:cover]{terra::cover()}} to fill
in missing values in source data with zero.

If \code{SDM = "riparian"}, pixel counts are converted to a proportion of the
total number of cells expected within the buffer distance represented by
\code{scale}, and the \code{scale} is appended to the predictor name in the format
"_50" or "_2000", as expected by the riparian SDMs.

If \code{SDM = "waterbird_fall"} or \code{SDM = "waterbird_win"}, the \code{scale} is
appended to the predictor name in the format "_2k", "_5k", or "_10k", as
expected by the waterbird SDMs.

The final rasters are then written to the directory \code{pathout/scenario},
which will be created if it doesn't yet exist.
}
\examples{
# See vignette

}
\seealso{
\code{\link[=python_focal_prep]{python_focal_prep()}}, \code{\link[=python_focal_run]{python_focal_run()}}
}
