% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_focal_prep.R
\name{python_focal_prep}
\alias{python_focal_prep}
\title{Prepare landscape rasters for focal statistics via Python}
\usage{
python_focal_prep(
  x,
  SDM,
  classified = FALSE,
  fill = TRUE,
  suffix = NULL,
  pixel_value = NULL,
  subset = NULL,
  dir = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{SpatRaster with layer names corresponding to \code{landscape_name}}

\item{SDM}{The name of intended species distribution model, for which \code{x}
will be reclassified: \code{"riparian"}, \code{"waterbird_fall"}, \code{"waterbird_win"},
or \code{"tima"}}

\item{classified}{logical; passed to \code{\link[=create_predictor_stack]{create_predictor_stack()}}}

\item{fill}{logical; passed to \code{\link[=create_predictor_stack]{create_predictor_stack()}}}

\item{suffix}{Character string; custom suffix appended to layer names
(optional unless \code{mask} is not \code{NULL}); see Details.}

\item{pixel_value}{Optional numeric value to use in place of 1 where land
covers are present; default \code{NULL}}

\item{subset}{Optional SpatRaster or string representing filepath to a
raster. See Details.}

\item{dir}{Optional string representing directory passed to
\code{\link[terra:writeRaster]{terra::writeRaster()}}, as (\code{dir/SDM/landscape_name}). See Details.}

\item{verbose}{logical; passed to \code{\link[=classify_landcover.SpatRaster]{classify_landcover.SpatRaster()}}}

\item{...}{additional arguments passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}}
}
\value{
SpatRaster, though primarily used to write layers to file for use
with \code{\link[=python_focal_stats]{python_focal_stats()}}
}
\description{
Prepare for running focal statistics on a landscape raster via Python, to
generate inputs for use with species distribution models.
}
\details{
This is a wrapper function that calls \code{\link[=create_predictor_stack]{create_predictor_stack()}} to split a
land cover raster into separate layers representing the presence (1) or
absence (0) of each required land cover class predictor for use in
calculating focal statistics. The input raster should already be encoded with
the land cover classes listed in the \link{key}. To help with creating such a
raster, see \code{\link[=classify_landcover.sf]{classify_landcover.sf()}} to map land cover polygons to the land
cover classes in the \link{key}. See documentation of \code{\link[=create_predictor_stack]{create_predictor_stack()}}
and \code{\link[=classify_landcover.SpatRaster]{classify_landcover.SpatRaster()}} for information about Warning messages.

If \code{suffix} is provided, it is appended to the layer name. If \code{pixel_value}
is provided, values representing land cover presence (1) are replaced with
this value (e.g., the area of each pixel).

If \code{subset} is provided, either as a SpatRaster or a filepath to a raster,
the \code{subset} layer is masked by the presence of each land cover layer, e.g.
to extract from surface water data the extent of surface water within each
land cover class. In this case, both the land cover presence data and the
subset data are returned; to distinguish between these results, \code{suffix} is
required to have two values. The first is appended to the land cover presence
data and the second is appended to the subset data. See examples.

All output can be optionally written to \code{dir/SDM/landscape_name} where
\code{landscape_name} is taken from the names of the input raster(s) \code{x}.
}
\examples{
codenums = DeltaMultipleBenefits::key$CODE_NUM
r <- terra::rast(matrix(sample(codenums, size = 1000, replace = TRUE), ncol = 100, nrow = 100))
watwin = suppressWarnings(classify_landcover(r, SDM = 'waterbird_win', verbose = FALSE))
watwin_pred = python_focal_prep(watwin, SDM = 'waterbird_win')

# return the area of the pixel where each land cover class is present
# (useful for summing over moving windows)
watwin_area = python_focal_prep(watwin, SDM = 'waterbird_win', pixel_value = 0.09)

# mask another raster (e.g., surface water data) by the presence of each
# land cover class:
w = watwin # simulate surface water data
levels(w) = NULL
terra::coltab(w) = NULL
terra::values(w) <- sample(c(0,1), size = 10000, replace = TRUE)
#pfld = python_focal_prep(watwin, SDM = 'waterbird_win', pixel_value = 0.09, mask = w)
#returns error because two suffixes need to be provided
pfld = python_focal_prep(watwin, SDM = 'waterbird_win', pixel_value = 0.09, mask = w,
                         suffix = c('_area', '_pfld')) # works
}
\seealso{
\code{\link[=python_focal_stats]{python_focal_stats()}}
}
