% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_focal_run.R
\name{python_focal_run}
\alias{python_focal_run}
\title{Run focal statistics via Python}
\usage{
python_focal_run(
  pathin,
  landscape_name,
  SDM,
  regex = NULL,
  scale,
  suffix,
  fun = "MEAN",
  dir,
  python = NULL
)
}
\arguments{
\item{pathin, SDM, landscape_name}{Character strings defining the filepath
(\code{pathin/SDM/landscape_name}) containing input rasters to be processed,
such as those written from running \code{\link[=python_focal_prep]{python_focal_prep()}}}

\item{regex}{Optional regular expression to process only a subset of the
rasters in \code{pathin/SDM/landscape_name}}

\item{scale}{String representing the buffer size (in m) within which focal
stats are calculated}

\item{fun}{Function to summarize focal statistics: \code{'MEAN'} or \code{'SUM'}}

\item{dir}{Filepath for the directory where output rasters should be written
(as \code{dir/SDM/landscape_name/scale})}

\item{python}{Optional filepath to the preferred version of arcpy, passed to
\code{reticulate::use_python}. See details.}
}
\value{
Nothing returned to R environment. Writes rasters to \code{pathout} for
each land cover class.
}
\description{
Internal function that checks for the availability of Python, imports the
arcpy module and Spatial Analyst extensions, then sources the Python script
"focal_stats.py".
}
\details{
This function is primarily for internal use by
\code{\link[=python_focal_stats]{python_focal_stats()}} to calculate focal statistics for each land cover
predictor, passing the appropriate moving window sizes and summary
functions required by each species distribution model. For each unique
directory of land cover predictors, this function calls the
\code{focal_stats.py} script which reads in all rasters in the directory (unless
the \code{regex} argument is used) and then summarizes cell values for each
input raster within a buffer distance defined by \code{scale}.

Summary functions may include \code{'SUM'} or \code{'MEAN'}. Note that the MEAN of
binary land cover presence data is equivalent to the proportion cover of
each land cover class within the buffer distance while the SUM represents
the count of pixels within the buffer distance.
}
\examples{
# See vignette
}
\seealso{
\code{\link[=python_focal_prep]{python_focal_prep()}}, \code{\link[=python_focal_finalize]{python_focal_finalize()}}
}
\keyword{internal}
