% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_focal_stats.R
\name{python_focal_stats}
\alias{python_focal_stats}
\title{Run focal statistics via Python}
\usage{
python_focal_stats(
  SDM,
  pathin,
  dir,
  landscape_names,
  regex = NULL,
  python = NULL,
  overwrite = TRUE,
  mask = NULL
)
}
\arguments{
\item{pathin, SDM, landscape_names}{Character strings defining the filepath
(\code{pathin/SDM/landscape_names}) containing input rasters to be processed,
such as those written from running \code{\link[=python_focal_prep]{python_focal_prep()}}}

\item{dir}{Filepath for the directory where output rasters should be written
(as \code{dir/SDM/landscape_name/scale})}

\item{regex}{Optional regular expression to process only a subset of the
rasters in \code{pathin/SDM/landscape_name}}

\item{python}{Optional filepath to the preferred version of arcpy, passed to
\code{reticulate::use_python}. See details.}

\item{overwrite}{logical; allow Python to overwrite existing output?}

\item{mask}{currently experimental}
}
\value{
None
}
\description{
Wrapper function to call a python script for calculating focal statistics on
land cover predictors for species distribution models. On first call this
function imports the arcpy module and Spatial Analyst extensions then sources
the Python script "focal_stats.py".
}
\details{
This function is designed to be called after running
\code{\link[=python_focal_prep]{python_focal_prep()}}, and writing the resulting rasters representing land
cover predictors to \code{pathin/SDM/landscape_names}. For each raster in these
directories, this function calculates focal statistics required for each
species distribution model, on the appropriate spatial scales and with the
required summary statistics. The \code{regex} argument provides options for
processing only a subset of the rasters in the directory. If
\code{overwrite=TRUE} (the default), previously-created focal statistics will be
overwritten. If \code{mask} provides a filepath to a raster, the results will be
masked by this file before writing to file.

These calculations can be very slow, depending on the size and resolution
of the rasters, and relies on the availability of \code{arcpy} and Spatial
Analyst extensions for faster processing. An attempt will be made to load
these the first time this function (or \code{\link[=python_dist]{python_dist()}}) is called in each
session, and by default will look here: \verb{C:/Program Files/ArcGIS/Pro/bin/Python/envs/arcgispro-py3/python.exe}; use the
\code{python} argument to specify a different pathway.
}
\examples{
# See vignette
}
\seealso{
\code{\link[=python_focal_prep]{python_focal_prep()}}, \code{\link[=fit_SDM]{fit_SDM()}}
}
