% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify_landcover.R
\name{reclassify_landcover}
\alias{reclassify_landcover}
\title{Reclassify landscape rasters by SDM group}
\usage{
reclassify_landcover(landscape, SDM)
}
\arguments{
\item{landscape}{SpatRaster created by \code{\link[terra:rast]{terra::rast()}}}

\item{SDM}{The name of intended species distribution model: \code{"riparian"},
\code{"waterbird_fall"}, \code{"waterbird_win"}, or \code{"tima"}}
}
\value{
SpatRaster with separate layers for each land cover class included as
a predictor in the selected SDM representing the presence (1) and absence
(0)
}
\description{
Prepare for fitting SDMs by reclassifying a landscape rasters according to
the classifications used by a specific set of species distribution models
(SDM).
}
\details{
Calls on internal datsets to crosswalk land cover classes encoded in the
provided landscape predictors expected by the selected set of SDMs. This
function is called by \code{\link[=python_focal_prep]{python_focal_prep()}} on a set of landscape rasters
generated by segregating an input landscape by class. It is not intended to
be called directly.

Warnings are given if land cover classes are present in the landscape but are
excluded from model consideration, which may be a problem particularly if
they represent a large portion of the landscape. In this case, review the
corresponding internal dataset (\code{\link[=predictors_riparian]{predictors_riparian()}},
\code{\link[=predictors_waterbird_fall]{predictors_waterbird_fall()}}, \code{\link[=predictors_waterbird_win]{predictors_waterbird_win()}}, or
\code{\link[=predictors_tima]{predictors_tima()}}) to understand whether the model is expecting more
specific land cover classes or subclasses; the provided landscape may need to
be adjusted first.

Warnings are also given if land cover classes are expected by the model but
are absent from the landscape. These land covers will be assumed absent from
the provided landscape and filled in with all 0 values, but these missing
land covers should be carefully reviewed to ensure they aren't excluded from
the landscape and model predictions unintentionally. If needed, the resulting
layers can be replaced manually before proceeding with \code{\link[=python_focal_run]{python_focal_run()}}.
}
\examples{
library(DeltaMultipleBenefits)
#data(key)
r <- terra::rast(matrix(sample(c(11,19,71,72,90), size = 100, replace = TRUE),
         ncol = 10, nrow = 10)
landcover_presence = suppressWarnings(reclassify_landcover(r, SDM = 'riparian'))
}
\seealso{
\code{\link[=key]{key()}}, \code{\link[=predictors_riparian]{predictors_riparian()}}, \code{\link[=predictors_waterbird_fall]{predictors_waterbird_fall()}},
\code{\link[=predictors_waterbird_win]{predictors_waterbird_win()}}, or \code{\link[=predictors_tima]{predictors_tima()}} for more details on
the list of recognized landcover classes and subclasses and how they
crosswalk to land cover predictors for each set of SDMs
}
