% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_change.R
\name{sum_change}
\alias{sum_change}
\title{Summarize net change between scenario and baseline landscapes}
\usage{
sum_change(dat, k = 2)
}
\arguments{
\item{dat}{tibble; see Details}

\item{k}{coverage factor; see Details}
}
\value{
tibble containing fields \code{BASELINE}, \code{SCENARIO}, \code{net_change}, and
optionally \code{net_change_se}, \code{U}, \code{lcl}, \code{ucl}, and \code{z}; see Details
}
\description{
Calculate the difference in the landscape-level metric totals between a
baseline landscape and one or more scenario landscapes
}
\details{
This function expects \code{dat} to contain the following fields:
\itemize{
\item \code{scenario}: a character field used to identify the name of the landscape
being examined; one of these must be called \code{'baseline'} and all others will
be treated as alternate scenarios for comparison with the baseline
\item either \code{area} or \code{SCORE_TOTAL}: numeric fields representing the total
landscape-level values for each metric, such as those produced by
\code{\link[=sum_landcover]{sum_landcover()}} or \code{\link[=sum_metrics]{sum_metrics()}}.
}

The scores for each scenario are aligned with corresponding scores for the
baseline landscape based on any other common fields (e.g. \code{ZONE},
\code{METRIC_CATEGORY}) and the net difference is calculated as the scenario
score minus the baseline score. The function returns the original baseline
and scenario total landscape scores for each metric and scenario (renamed as
\code{BASELINE} and \code{SCENARIO}), along with \code{net_change}.

If \code{SCORE_TOTAL_SE} is also provided in \code{dat}, representing the uncertainty
in the \code{SCORE_TOTAL}, the uncertainty in the difference (\code{net_change_se}) is
also calculated as: \code{sqrt(baseline_se^2 + scenario_se^2)} where
\code{baseline_se} and \code{scenario_se} represent the \code{SCORE_TOTAL_SE} for the
baseline and scenario landscapes, respectively. In addition, the coverage
factor \code{k} is used to estimate expanded uncertainty (\code{U}), or the interval
within which a large fraction of the distribution of values could be
reasonably expected. The appropriate value for \code{k} depends on the level of
confidence required, the number of observations on which the uncertainty is
based, and any knowledge of the underlying distributions of the estimates.
Where the distributions concerned are normal, and for most purposes, a value
of 2 is recommended  to give an interval containing approximately 95\% of the
distribution of values. In this case, the function also returns the original
uncertainty estimates for the baseline and scenario (renamed as
\code{BASELINE_SE} and \code{SCENARIO_SE}), \code{net_change_se}, the expanded uncertainty
estimate \code{U} and corresponding \code{lcl} and \code{ucl}, as well as a \code{z} score equal
to \code{abs(net_change/net_change_se)}.
}
\examples{
# See vignette
}
\seealso{
\code{\link[=sum_habitat]{sum_habitat()}}, \code{\link[=sum_metrics]{sum_metrics()}}
}
