% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_habitat.R
\name{sum_habitat}
\alias{sum_habitat}
\title{Summarize total habitat scores}
\usage{
sum_habitat(
  pathin,
  zonepath = NULL,
  subtype = "distributions",
  rollup = TRUE,
  keypath = NULL,
  scale = NULL
)
}
\arguments{
\item{pathin}{Character string defining the filepath to the highest-level
directory containing the predicted presence/absence or probability of
presence from species distribution models, such as those created from
running \code{\link[=fit_SDM]{fit_SDM()}} or \code{\link[=transform_SDM]{transform_SDM()}}}

\item{zonepath}{Optional character string defining the filepath to a raster
encoding zones within which pixel values should be summarized}

\item{subtype}{Character string appended to the field METRIC_SUBTYPE, such as
for distinguishing probability of presence from presence/absence}

\item{rollup}{Logical; If \code{TRUE} (default), summarize total habitat across
all species/groups by set of SDMs}

\item{keypath}{Optional filepath passed to \code{\link[readr:read_delim]{readr::read_csv()}} translating
the individual species names as encoded in the file names in pathin to
METRIC names in the output table}

\item{scale}{Optional value by which to scale the results; see Details}
}
\value{
tibble
}
\description{
Calculate the sum of the predicted presence/absence or probability of
presence from species distribution models corresponding to a set of landscape
rasters.
}
\details{
This function will calculate the sum of all pixel values in each
raster found in \code{pathin} and any subdirectories. The file structure within
\code{pathin} is used to infer the name of the corresponding \code{SDM} and
\code{landscape_name}. This function is designed to efficiently process multiple
rasters located in the \code{pathin} directory at once.

If provided, \code{keypath} should be a CSV containing the fields \code{spp} and
\code{label}, used for converting the filenames of the rasters within \code{pathin}
to a more readable label, which will be renamed \code{METRIC} in the final
output.

If provided, the sum of all pixel values will be multiplied by \code{scale}. For
example, to rescale the total in terms of the total area, enter the area of
each pixel.
}
\examples{
# See vignette
}
\seealso{
\code{\link[=fit_SDM]{fit_SDM()}}, \code{\link[=transform_SDM]{transform_SDM()}}
}
