% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_landcover.R
\name{sum_landcover}
\alias{sum_landcover}
\title{Summarize total area of land cover classes}
\usage{
sum_landcover(pathin, maskpath, pixel_area = 1, zonepath = NULL, rollup = TRUE)
}
\arguments{
\item{pathin}{Character string defining the filepath to a directory
containing landscapes (e.g. scenarios) to be summarized.}

\item{maskpath}{Optional filepath to a raster that should be used to mask the
output, e.g. a study area boundary}

\item{pixel_area}{Numeric value representing the area of each pixel, in any
units; will be multiplied by the number of pixels of each land cover class}

\item{zonepath}{Optional character string defining the filepath to a raster
encoding zones within which pixel values should be summarized}

\item{rollup}{Logical; whether or not riparian and managed wetland subclasses
should be rolled up into a total area of riparian and managed wetlands,
respectively.}
}
\value{
tibble
}
\description{
Summarize the area or number of pixels of each land cover class from a set of
landscape rasters
}
\details{
This function will summarize the frequency of pixels of each land
cover class in each landscape raster provided in \code{pathin}, optionally
masked by a raster at \code{maskpath}, and/or optionally summarized by zones
encoded in a raster at \code{zonepath}. The resulting number of pixels is
multiplied by \code{pixel_area} as an estimate of the total area of each land
cover class in each landscape. This function is designed to efficiently
process several landscapes located in the \code{pathin} directory at once.

This function is generally agnostic as to the names of the land cover
classes it expects, but it does expect the landscapes to have levels
defined (see \code{\link[terra:factors]{terra::levels()}}) and not just contained numeric codes. One
exception is if \code{rollup = TRUE}, in which case it will expect to find
riparian and managed wetland subclasses with names that begin with
\code{'RIPARIAN_'} or \code{'WETLAND_MANAGED_'}. These subclasses will be summarized
into a total area of \code{'RIPARIAN'} and \code{'WETLAND_MANAGED'} land cover
classes.

\code{pixel_area} can be provided in any units you prefer, but if this function
is intended to produce \code{areadat} for use with \code{\link[=sum_metrics]{sum_metrics()}}, take care
that these units align with the UNITs by which \code{metricdat} are defined
}
\examples{
# See vignette
}
\seealso{
\code{\link[=sum_habitat]{sum_habitat()}}, \code{\link[=sum_metrics]{sum_metrics()}}
}
