% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_SDM.R
\name{transform_SDM}
\alias{transform_SDM}
\title{Transform predictions from species distribution models to binary}
\usage{
transform_SDM(pathin, landscape_name, modist, stat, pathout, overwrite = FALSE)
}
\arguments{
\item{pathin, landscape_name}{Character strings defining the filepath
(\code{pathin/landscape_name}) containing the predicted probability of presence
resulting from each distribution model, such as those created from running
\code{\link[=fit_SDM]{fit_SDM()}}}

\item{stat}{Character string defining the threshold statistic to be used; see
\code{\link[dismo:threshold]{dismo::threshold()}} for options}

\item{pathout}{Filepath for the directory where results rasters should be
written}

\item{overwrite}{Logical; passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}}

\item{modlist}{List of model objects of class 'gbm' representing the
distribution models to which new predictors should be fit.}
}
\value{
Nothing returned to R environment. Writes rasters to \code{pathout} for
each model in \code{modlist}
}
\description{
Use model-specific threshold values to transform predicted probabilities of
species presence, such as resulting from \code{\link[=fit_SDM]{fit_SDM()}}, to binary predictions
of presence or absence.
}
\examples{
# See vignette
}
\seealso{
\code{\link[=fit_SDM]{fit_SDM()}}
}
