% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_covertype.R
\name{update_covertype}
\alias{update_covertype}
\title{Update waterbird and tidal marsh bird predictors: covertype and LANDCOVER}
\usage{
update_covertype(
  landscape,
  key,
  SDM,
  mask = NULL,
  pathout,
  landscape_name,
  overwrite = FALSE
)
}
\arguments{
\item{landscape}{SpatRaster created by \code{\link[terra:rast]{terra::rast()}}}

\item{key}{tibble, dataframe, or character string defining a filepath passed
to \code{\link[readr:read_delim]{readr::read_csv()}}, used to interpret the raster values in \code{landscape}
to land cover class names; see Details}

\item{SDM}{The name of intended species distribution model:
\code{"waterbird_fall"}, \code{"waterbird_win"}, or \code{"tima"}}

\item{mask}{Optional filepath to a raster that should be used to mask the
output, e.g. a study area boundary}

\item{pathout, landscape_name}{Character strings defining the filepath
(\code{pathout/SDM/landscape_name}) where output rasters should be written}

\item{overwrite}{Logical; passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}; default \code{FALSE}}
}
\value{
Nothing; all files written to \code{pathout}
}
\description{
Helper function for updating covertype and LANDCOVER predictors for the
waterbird and tidal marsh bird distribution models, respectively.
}
\details{
Classifies the \code{landscape} rasters according to the land cover
classes that were originally surveyed, which are the only classes for which
predictions should be generated from the waterbird distribution models. For
\code{"waterbird_fall"} or \code{"waterbird_win"} SDMs, generates file
\code{covertype.tif} and for \code{"tima"} SDMs, generates file \code{LANDCOVER.tif}.
Output is written to \verb{pathout/SDM/scenario_name/}.
}
\examples{
# See vignette
}
\seealso{
\code{\link[=update_pwater]{update_pwater()}}; \code{\link[=update_roosts]{update_roosts()}}
}
