% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_covertype.R
\name{update_covertype}
\alias{update_covertype}
\title{Update waterbird and tidal marsh bird predictors: covertype and LANDCOVER}
\usage{
update_covertype(x, SDM, mask = NULL, pathout = NULL, overwrite = FALSE)
}
\arguments{
\item{x}{SpatRaster}

\item{SDM}{The name of intended species distribution model:
\code{"waterbird_fall"}, \code{"waterbird_win"}, or \code{"tima"}}

\item{mask}{Optional SpatRaster or string representing filepath to a raster
that should be used to mask the output, e.g. a study area boundary}

\item{overwrite}{logical. If \code{TRUE}, output is overwritten}

\item{dir}{Optional string representing output directory
(\code{dir/SDM/landscape_name}) where output rasters should be written}
}
\value{
SpatRaster with the same number of layers as the input \code{x}, names
required by the selected \code{SDM}.
}
\description{
Helper function for updating \code{covertype} and \code{LANDCOVER} predictors for the
waterbird ("waterbird_fall", "waterbird_win") and tidal marsh ("tima")
distribution models, respectively.
}
\details{
Classifies land cover rasters to generate categorical predictors
required by the "waterbird_fall", "waterbird_win", and "tima" models. The
input raster \code{x} should already represent land cover predictors required by
the selected \code{SDM}, i.e. output from \code{\link[=classify_landcover.SpatRaster]{classify_landcover.SpatRaster()}}. A
mask raster can optionally be provided, either as a SpatRaster or a
filepath to a raster, to first mask the input raster(s) \code{x}, . Output can
be optionally written to \code{dir/SDM/landscape_name} where \code{landscape_name} is
taken from the names of the input raster(s) \code{x}.
}
\examples{
# See vignette
}
\seealso{
\code{\link[=classify_landcover.SpatRaster]{classify_landcover.SpatRaster()}}; \code{\link[=update_pwater]{update_pwater()}};
\code{\link[=update_roosts]{update_roosts()}}
}
