% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_covertype.R
\name{update_covertype}
\alias{update_covertype}
\title{Update waterbird predictors: covertype}
\usage{
update_covertype(
  landscape,
  SDM,
  maskpath = NULL,
  pathout,
  scenario_name,
  overwrite = FALSE
)
}
\arguments{
\item{landscape}{SpatRaster created by \code{\link[terra:rast]{terra::rast()}}}

\item{SDM}{The name of intended species distribution model:
\code{"waterbird_fall"}, or \code{"waterbird_win"}}

\item{maskpath}{Optional filepath to a raster that should be used to mask the
output, e.g. a study area boundary}

\item{pathout}{Character string; Filepath to directory where output rasters
should be written; passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}}

\item{scenario_name}{Character string; Name of the landscape scenario being
evaluated, corresponding to the directory in \code{pathout} where results will
be written.}

\item{overwrite}{Logical; passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}; default \code{FALSE}}
}
\value{
Nothing; all files written to \code{pathout}
}
\description{
Helper function for updating covertype predictors for the waterbird
distribution models.
}
\details{
Classifies the landscape rasters according to the land cover classes
that were originally surveyed (the only classes for which predictions should
be generated). Generates file: \code{covertype.tif}.
}
\examples{
# See vignette
}
\seealso{
\code{\link[=update_pwater]{update_pwater()}}; \code{\link[=update_roosts]{update_roosts()}}
}
