% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pwater.R
\name{update_pwater}
\alias{update_pwater}
\title{Update waterbird predictors: pwater}
\usage{
update_pwater(
  landscape,
  scenario_landscape = NULL,
  pathout,
  scenario_name = NULL,
  overwrite = FALSE,
  waterdatpath,
  maskpath,
  floor = TRUE
)
}
\arguments{
\item{landscape}{SpatRaster created by \code{\link[terra:rast]{terra::rast()}}}

\item{scenario_landscape}{Optional secondary SpatRaster created by
\code{\link[terra:rast]{terra::rast()}} to compare against landscape in \code{\link[=generate_pwater]{generate_pwater()}}; see
Details.}

\item{pathout}{Character string; Filepath to directory where output rasters
should be written; passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}}

\item{scenario_name}{Character string; Name of the landscape scenario being
evaluated, corresponding to the directory in \code{pathout} where results will
be written.}

\item{overwrite}{Logical; passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}; default \code{FALSE}}

\item{waterdatpath}{Filepath to a raster representing the probability of open
water (pwater) in each cell, specific to the time frames for each waterbird
SDM}

\item{maskpath}{Optional filepath to a raster that should be used to mask the
output, e.g. a study area boundary}

\item{floor}{Logical; if \code{TRUE}, don't allow new values of pwater to be lower
than baseline values}
}
\value{
Nothing; all files written to \code{pathout}
}
\description{
Helper function for updating pwater predictors for the waterbird distribution
models.
}
\details{
Uses surface water data in \code{waterdatpath} to extract cell-specific
probabilities of open water for the \code{baseline_landscape}; if
\code{scenario_landscape} is also provided, representing a projected change from
the baseline, assign new cell-specific probabilities of surface water for
the cells that have changed class, based on the mean probability of open
water for each land cover class in the baseline landscape. If \code{floor = TRUE}, new probabilities of open water will be assigned only if they are
highthan the baseline values. Generates file: \verb{pwater.tif'}. Use this
function prior to using \code{\link[=python_focal_prep]{python_focal_prep()}} with \code{mask} = 'pwater.tif' to
generate probability of flooding data by land cover class for use with the
waterbird SDMs.
}
\examples{
# See vignette
}
\seealso{
\code{\link[=update_covertype]{update_covertype()}}, \code{\link[=update_roosts]{update_roosts()}}
}
