% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pwater.R
\name{update_pwater}
\alias{update_pwater}
\title{Update waterbird predictors: pwater & pfld}
\usage{
update_pwater(
  waterdat,
  mask = NULL,
  pathout,
  SDM,
  landscape_name,
  overwrite = FALSE,
  baseline_landscape = NULL,
  scenario_landscape = NULL,
  floor = FALSE
)
}
\arguments{
\item{waterdat}{\code{SpatRaster} or character string giving the filepath to a
raster representing the probability of open water (pwater) in each cell,
specific to the time frames appropriate to each \code{scenario_landscape} and
waterbird SDM (i.e., fall vs. winter)}

\item{mask}{Optional \code{SpatRaster} or character string giving the filepath to
a raster that should be used to mask the output, e.g. a study area boundary}

\item{pathout, SDM, landscape_name}{Character strings defining the filepath
(\code{pathout/SDM/landscape_name}) where output rasters should be written;
landscape_name should either correspond to the landscape represented by
\code{waterdat} or the \code{scenario_landscape}, if given; see Details}

\item{overwrite}{Logical; passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}; default \code{FALSE}}

\item{baseline_landscape, scenario_landscape}{Optional SpatRasters created by
\code{\link[terra:rast]{terra::rast()}} to compare with each other for estimating \code{pwater} for the
changed portions of the \code{scenario_landscape}; see Details}

\item{floor}{Logical; if \code{TRUE}, don't allow new values of pwater to be lower
than baseline values}
}
\value{
Nothing; all files written to \code{pathout}
}
\description{
Helper function for updating \code{pwater} and \code{pfld} predictors for the waterbird
distribution models. Generates file \code{pwater.tif} at locations
\code{pathout/pwater/landscape_name} and \code{pathout/SDM/landscape_name}.
}
\details{
The waterbird distribution models incorporate information about
surface water data in two ways: as \code{pwater}, the expected probability of
open surface water in each cell of the landscape raster, specific to the
waterbird season being modeled and perhaps averaged over multiple years,
and as \code{pfld} focal statistics which represent the proportion of each land
cover class within a given distance of each cell that is flooded (see
\code{\link[=python_focal_prep]{python_focal_prep()}} and \code{\link[=python_focal_run]{python_focal_run()}}). Therefore, \code{pwater} data
must be available for every landscape under analysis before the \code{pfld}
focal statistics can be generated and distribution models fit.

Due to the dual needs for generating \code{pwater} and \code{pfld} predictors, this
function writes results in two places within \code{pathout}. The first will be
written to \code{pathout/pwater/landscape_name}, intended for later use with
\code{\link[=python_focal_prep]{python_focal_prep()}} and generating \code{pfld} predictors. The second will be
written to \code{pathout/SDM/landscape_name}, which is expected to be a
directory containing all final predictors for later use with \code{\link[=fit_SDM]{fit_SDM()}} in
fitting waterbird models.

In addition, this function has two modes of operation. If
\code{scenario_landscape} is not provided, the \code{waterdat} is assumed to to
represent \code{pwater} data for the \code{landscape_name}, and is simply renamed and
copied to both \code{pathout} locations for use in later steps of analysis,
optionally masking before \code{pathout/SDM/landscape_name} is written. The
\code{mask} is never applied to the \code{pathout/pwater/landscape_name} output
intended for later focal statistics to avoid errors in processing near the
boundaries of the study area.

Alternatively, in the second mode, if both \code{baseline_landscape} and
\code{scenario_landscape} rasters are provided, this function will estimate
new \code{pwater} values for cells in the \code{scenario_landscape} that have changed
cover class, based on the mean probability of open surface water for that
land cover class in the \code{baseline_landscape}. Optionally, if \code{floor = TRUE}, new probabilities of open water will be assigned only if they are
higher than the baseline values. In this mode, the result represents
\code{pwater} for the \code{scenario_landscape}, and \code{landscape_name} should reflect
the name of the scenario.

The original \code{pwater} baseline data used in the development of these models
was derived from Point Blue's \href{https://www.pointblue.org/autowater}{Water Tracker} and may be downloaded from
\link{link TBD}.
}
\examples{
# See vignette
}
\seealso{
\code{\link[=update_covertype]{update_covertype()}}, \code{\link[=update_roosts]{update_roosts()}}
}
