% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_roosts.R
\name{update_roosts}
\alias{update_roosts}
\title{Update waterbird predictors: crane roost locations}
\usage{
update_roosts(
  x,
  unsuitable = c(10:19, 60, 70:79, 100:120, 170:187),
  proportion = 0.2,
  roosts,
  filename = NULL,
  ...
)
}
\arguments{
\item{x}{SpatRaster}

\item{unsuitable}{numerical vector representing the land cover
classifications in \code{x} that are incompatible with crane roosts}

\item{proportion}{numerical threshold at which cover by an unsuitable land
cover class makes a roost polygon unsuitable; see Details}

\item{roosts}{SpatVector or character string giving the filepath to polygons
representing the location of traditional crane roosts}

\item{filename}{Optional character string passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}}

\item{...}{Additional arguments passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}}
}
\description{
Helper function for estimating impact of landscape changes on the locations
of known crane roosts.
}
\details{
For landscapes that represent a projected change from baseline
conditions, this function facilitates evaluating historical crane roost
polygons to determine whether the land cover overlaying them in raster \code{x}
is now unsuitable, based on exceeding a threshold \code{proportion} covered by
an unsuitable land cover class. Unsuitable roost polygons are removed and
the rest rasterized to match \code{x}. Use this function prior to using
\code{\link[=python_dist]{python_dist()}} to calculate distance to roost and generate updated
versions of \code{droost_km.tif} for each scenario.

The default values for \code{unsuitable} include the original encodings for
orchard and vineyard classes (10-19), urban (60), riparian (70-79,
170-187), and woodland and scrub (100-120), and the default threshold value
for \code{proportion} is 0.2.
}
\examples{
# See vignette
}
\seealso{
\code{\link[=update_covertype]{update_covertype()}}, \code{\link[=update_pwater]{update_pwater()}}
}
