% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_roosts.R
\name{update_roosts}
\alias{update_roosts}
\title{Update waterbird predictors: crane roost locations}
\usage{
update_roosts(landscape, roostpath, pathout, scenario_name, overwrite = FALSE)
}
\arguments{
\item{landscape}{SpatRaster created by \code{\link[terra:rast]{terra::rast()}}}

\item{roostpath}{Filepath to a raster representing the location of
traditional crane roosts.}

\item{pathout}{Character string; Filepath to directory where output rasters
should be written; passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}}

\item{scenario_name}{Character string; Name of the landscape scenario being
evaluated, corresponding to the directory in \code{pathout} where results will
be written.}

\item{overwrite}{Logical; passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}; default \code{FALSE}}
}
\value{
Nothing; all files written to \code{pathout}
}
\description{
Helper function for estimating impact of landscape changes on known crane
roosts, to generate updated estimates of the distance to roost.
}
\details{
For landscapes that represent a projected change from baseline conditions,
evaluates historical crane roosts to determine whether the land cover is
projected to become unsuitable (>20\% cover with an unsuitable land cover
class). Unsuitable roosts are removed, generating files named for
\code{scenario_name} in \code{pathout}. Use this function prior to using \link{python_dist}
to calculate distance to roost and generate \code{droost_km.tif}.
}
\examples{
# See vignette
}
\seealso{
\code{\link[=update_covertype]{update_covertype()}}, \code{\link[=update_pwater]{update_pwater()}}
}
