% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_roosts.R
\name{update_roosts}
\alias{update_roosts}
\title{Update waterbird predictors: crane roost locations}
\usage{
update_roosts(
  landscape,
  unsuitable = c(11:19, 60, 70:79, 100:120),
  proportion = 0.2,
  roostpath,
  pathout,
  landscape_name,
  overwrite = FALSE
)
}
\arguments{
\item{landscape}{SpatRaster created by \code{\link[terra:rast]{terra::rast()}}}

\item{unsuitable}{optional vector of numerical values representing the land
cover classifications that should be considered incompatible with crane
roosts; default values for the original land cover encoding include
orchard, vineyard, riparian, woodland, scrub, urban}

\item{proportion}{numerical value for the proportion cover by an unsuitable
land cover class at which the roost should be considered unsuitable; see
Details}

\item{roostpath}{Filepath to a raster representing the location of
traditional crane roosts.}

\item{pathout}{Character string; Filepath to directory where output rasters
should be written; passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}}

\item{landscape_name}{Character string; Name of the landscape being
evaluated, corresponding to the directory in \code{pathout} where results will
be written.}

\item{overwrite}{Logical; passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}; default \code{FALSE}}
}
\value{
Nothing; all files written to \code{pathout}
}
\description{
Helper function for estimating impact of landscape changes on known crane
roosts, to generate updated estimates of the distance to roost for use with
waterbird distribution models.
}
\details{
For landscapes that represent a projected change from baseline
conditions, this function facilitates evaluating historical crane roosts to
determine whether the land cover overlaying them in \code{landscape} is
projected to become unsuitable, based on exceeding a threshold \code{proportion}
covered by an unsuitable land cover class. Unsuitable roost polygons are
removed, and updated roost maps are generated, named as \code{scenario_name} in
\code{pathout}. Use this function prior to using \link{python_dist} to calculate
distance to roost and generate updated versions of \code{droost_km.tif} for each
scenario.

The default values for \code{unsuitable} include the original encodings for
orchard, vineyard, riparian, woodland, scrub, and urban, and the default
threshold value for \code{proportion} is 0.2. Alternate values can be provided
as desired.
}
\examples{
# See vignette
}
\seealso{
\code{\link[=update_covertype]{update_covertype()}}, \code{\link[=update_pwater]{update_pwater()}}
}
