% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_roosts.R
\name{update_roosts}
\alias{update_roosts}
\title{Update waterbird predictors: crane roost locations}
\usage{
update_roosts(landscape, roostpath, pathout, landscape_name, overwrite = FALSE)
}
\arguments{
\item{landscape}{SpatRaster created by \code{\link[terra:rast]{terra::rast()}}}

\item{roostpath}{Filepath to a raster representing the location of
traditional crane roosts.}

\item{pathout}{Character string; Filepath to directory where output rasters
should be written; passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}}

\item{landscape_name}{Character string; Name of the landscape being
evaluated, corresponding to the directory in \code{pathout} where results will
be written.}

\item{overwrite}{Logical; passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}; default \code{FALSE}}
}
\value{
Nothing; all files written to \code{pathout}
}
\description{
Helper function for estimating impact of landscape changes on known crane
roosts, to generate updated estimates of the distance to roost for use with
waterbird distribution models.
}
\details{
For landscapes that represent a projected change from baseline
conditions, this function facilitates evaluating historical crane roosts to
determine whether the land cover overlaying them in \code{landscape} is
projected to become unsuitable (>20\% cover with an unsuitable land cover
class, including orchard, vineyard, riparian, woodland, scrub, urban).
Unsuitable roost polygons are removed, and updated roost maps are
generated, named as \code{scenario_name} in \code{pathout}. Use this function prior
to using \link{python_dist} to calculate distance to roost and generate updated
versions of \code{droost_km.tif} for each scenario.
}
\examples{
# See vignette
}
\seealso{
\code{\link[=update_covertype]{update_covertype()}}, \code{\link[=update_pwater]{update_pwater()}}
}
